feat: Use dbus api to listen for unit events
Use the dbus api to listen for unit events. This simplifies things by
not having systemd overrides as a dependency for calling back into the
charm (which can be inconvenient if the charm is doing overrides as
well). This also has the side effect of being more responsive to the
systemd service ActiveState changes since there's a notification.

Signed-off-by: Billy Olsen <billy.olsen@canonical.com>